add users-lookup on users child_added
It is necessary to know where each duplicate data entry exists. To track
this, the users-lookup path is being use to store encoded strings that
point to each path that a simplified user object is stored.

Possible improvement: store a path to each variable of user data instead
of the entire simplified object. This would take the guess work out of
how much of the user data needs to be saved at each path.